Refactor: Change name of folder
