+++ src/auth/auth.resolver.ts
@@ -4,8 +4,8 @@ import { AuthService } from './auth.service';
 import { SignInInput, SignUpInput } from 'src/auth/inputs/auth.input';
 import { UseGuards } from '@nestjs/common';
 import { User } from 'src/user/entities/user.entity';
-import { CurrentUser } from 'src/modules/decorators/user.decorator';
-import { SignInGuard } from 'src/modules/guards/graphql-signin-guard';
+import { CurrentUser } from 'src/common/decorators/user.decorator';
+import { SignInGuard } from 'src/common/guards/graphql-signin-guard';
 
 @Resolver()
 export class AuthResolver {
