+++ src/common/modules/typeorm.module.ts
@@ -1,7 +1,7 @@
 import { DataSource } from 'typeorm';
 import { DynamicModule, Provider } from '@nestjs/common';
 import { getDataSourceToken } from '@nestjs/typeorm';
-import { TYPEORM_CUSTOM_REPOSITORY } from './typeorm.decorator';
+import { TYPEORM_CUSTOM_REPOSITORY } from '../decorators/typeorm.decorator';
 
 export class TypeOrmExModule {
   public static forCustomRepository<T extends new (...args: unknown[]) => any>(
