+++ src/place/place.repositoy.ts
@@ -1,4 +1,4 @@
-import { CustomRepository } from '../modules/decorators/typeorm.decorator';
+import { CustomRepository } from '../common/decorators/typeorm.decorator';
 import { Place } from './entities/place.entity';
 import { ExtendedRepository } from 'src/declare/declare.module';
