+++ src/user/user.resolver.ts
@@ -1,5 +1,5 @@
-import { CurrentUser } from '../modules/decorators/user.decorator';
-import { GraphqlPassportAuthGuard } from '../modules/guards/graphql-passport-auth.guard';
+import { CurrentUser } from '../common/decorators/user.decorator';
+import { GraphqlPassportAuthGuard } from '../common/guards/graphql-passport-auth.guard';
 import { UseGuards } from '@nestjs/common';
 import { Args, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { UserService } from './user.service';
@@ -7,7 +7,7 @@ import GraphQLJSON from 'graphql-type-json';
 import { GetManyInput, GetOneInput } from 'src/declare/inputs/custom.input';
 import { GetUserType, User } from './entities/user.entity';
 import { CreateUserInput, UpdateUserInput } from './inputs/user.input';
-import { CurrentQuery } from 'src/modules/decorators/query.decorator';
+import { CurrentQuery } from 'src/common/decorators/query.decorator';
 
 @Resolver()
 export class UserResolver {
